requirements.txt
suds==1.0.0
django-oscar==3.1
Django>=3.2
